Connectors: Update Tokens app connector
