+++ packages/connect-tokens/src/index.ts
@@ -1,5 +1,5 @@
-export { default as Token } from './entities/Token'
-export { default as TokenHolder } from './entities/TokenHolder'
-export { default as Tokens } from './entities/Tokens'
+export { default as Token } from './models/Token'
+export { default as TokenHolder } from './models/TokenHolder'
+export { default as Tokens } from './models/Tokens'
 export { default as TokensConnectorTheGraph } from './thegraph/connector'
 export { default } from './connect'
